refactor(commands): move eliteskyblock API calls from Hypixel API and into !weight command
The API is not related to Hypixel official API after all. Since it is not used anywhere else, it will be directly hosted on the command class.
