+++ src/core/hypixel/hypixel-skyblock-types.d.ts
@@ -322,10 +322,6 @@ export interface GardenResponse extends HypixelSuccessResponse {
   garden: Garden | undefined
 }
 
-export interface FarmingWeightResponse {
-  totalWeight: number
-}
-
 export interface Garden {
   garden_experience: number
   commission_data: { completed: Record<string, number> }
